Add a Nix setup allowing Nix users to invoke a Ruby/Jekyll setup
The users would have to run the nix-setup-or-initialize.sh file, which
will fetch Ruby and the dependencies from the Nix store. When that is
complete it is possible to run the Jekyll setup by running `nix-shell`
in the project directory.

Not everyone in the world is using Nix (yet), but this makes it very
easy to get started, and the Nix users will not have Ruby or Jekyll
installed globally on their systems, only for the Nix session.

I have ignored the .bundle folder as well; I do not know if this is
the way to go.

The setup is based on the following blog post by Nathan Bijnens:

  - https://nathan.gs/2019/04/19/using-jekyll-and-nix-to-blog/